requirements.txt
DateTime==4.4
pandas==1.4.1
psycopg2==2.9.3
python-decouple==3.6
requests==2.27.1
SQLAlchemy==1.4.32
virtualenv==20.4.3